langchain_doc_extractor/requirements.txt
langchain
langchain-openai
langchain-community
openai
tiktoken
PyMuPDF
fastapi
uvicorn
pydantic
python-multipart
pytesseract
pymupdf
Pillow
python-dotenv
